chore: alter int16 to uint256

+++ lesson-1/lib.sol
@@ -2,20 +2,20 @@
 pragma solidity ^0.8.16;
 
 library MathSafe{
-  function add(int16 a, int16 b)public pure returns(int16){
+  function add(uint256 a, uint256 b)public pure returns(uint256){
     return (a + b);
   }
 
-  function mul(int16 a, int16 b)public pure returns(int16){
+  function mul(uint256 a, uint256 b)public pure returns(uint256){
     return (a * b);
   }
 
-  function div(int16 a, int16 b)public pure returns(int16){
+  function div(uint256 a, uint256 b)public pure returns(uint256){
     require(b==0);
     return a/b; 
   }
 
-  function mod(int16 a, int16 b)public pure returns(int16){
+  function mod(uint256 a, uint256 b)public pure returns(uint256){
     require(b==0);
     return (a%b);
   }
